completed key application and update on README

+++ p_accumulate/custom_accumulator.cpp
@@ -4,14 +4,14 @@
 #include<vector>
 #include<thread>
 
-long accumulation_wrapper(std::vector<long>::iterator begin, std::vector<long>::iterator end, long& ref_val);
+long long accumulation_wrapper(std::vector<long>::iterator begin, std::vector<long>::iterator end, long long& ref_val);
 
-long CustomAccumulator::parallel(std::vector<long>::iterator begin,std::vector<long>::iterator end) {
+long long CustomAccumulator::parallel(std::vector<long>::iterator begin,std::vector<long>::iterator end) {
     if (end-begin < 1000) {
         return std::accumulate(begin, end, 0);
     }
     const long MAX_THREADS = 7;
-    std::vector<long> accumulation_values(MAX_THREADS, (long) 0);
+    std::vector<long long> accumulation_values(MAX_THREADS, (long) 0);
 
     int total_size = std::distance(begin, end);
     int average_accumulation_per_thread = total_size/MAX_THREADS;
@@ -30,7 +30,8 @@ long CustomAccumulator::parallel(std::vector<long>::iterator begin,std::vector<l
     for(auto& thread : thread_containers) {
         thread.join();
     }
-    long total_sum = std::accumulate(accumulation_values.begin(), accumulation_values.end(), 0);
+    long long total_sum = std::accumulate(accumulation_values.begin(), accumulation_values.end(), 0ll);
+
     for (int i = 0; i < MAX_THREADS; ++i) {
         std::cout << "accumulation val: " << accumulation_values[i] << std::endl;
     }
@@ -38,7 +39,7 @@ long CustomAccumulator::parallel(std::vector<long>::iterator begin,std::vector<l
     return total_sum;
 }
 
-long accumulation_wrapper(std::vector<long>::iterator begin,std::vector<long>::iterator end, long& ref_val) {
-    ref_val = std::accumulate(begin, end, (long) 0);
+long long accumulation_wrapper(std::vector<long>::iterator begin,std::vector<long>::iterator end, long long& ref_val) {
+    ref_val = std::accumulate(begin, end, 0ll);
     return ref_val;
 }
